requirements.txt
pandas
sqlalchemy
pymongo
python-dotenv
psycopg2-binary